requirements.txt
langchain>=1.2.0
langchain-openai>=0.2.0 
langchain-community>=0.3.0 
langchain-mcp-adapters>=0.1.0 
pydantic>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.0.0